backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
psycopg2-binary
bcrypt==3.2.0